Move useful stuff to a conftest.
